Add responsive table filtering with HTMX partial swapping
The original script was replaced with the use of HTMX to not have a
re-render happen during filtering. Now the filtering happens "inline"
through server-side HTML partial swappings.

This commit also introduces some template include tags for easier
reading and management. The use of include tags is subject to change if
their use turns out to be more cumbersome than direct HTML.